// src/entities/cities-modal-page-404/cities-modal-page-404.types.ts
import { CSSProperties } from 'react';
import { ICurrentCity } from 'widgets/page-404-shops/page-404-shops.types';

export interface CitiesListItemProps {
  name: string;
  cb: () => void;
}
export interface CitiesModalProps {
  label?: string;
  labelStyle?: CSSProperties;
  callback?: (city: ICurrentCity) => void;
}